Moved the basic CCR controls into C++
it is supposed to be all templates, but it does bloat the example. The
program binary size went from 500B to 630B. Ti compiler cl430 runs at
-O2 by default, and max is -O4. But with -O4 it gets only to 628B.

So, the templated static structs do not remain in compile-time only.

And now I need a disassembler to figure out what C++ gets wrong here and
how to nudge it to stay only in compile time.

Maybe it's just the compiler? It generates some code that is never used?

The templated static structs should be just a way to bundle functions.
I.e. it's just a nicely packaged C interface.

	modified:   reg_funcs.hpp