UP a couple more tries
So, this compile-time C++ code does not make sense for MSP430, because
Ti compiler cl430 does not really understand these features. It does not
inline and does not eliminate these templates of singleton structs with
constexpr static inline methods. They bloat the program .text memory.
And surely waste more cycles too.

I would like to use clang to compile this stuff. For the efficiency at
compile-time, and of course for C++17 or newer, with its constexpr. But
my main priority is Ti's features. I.e. I'd like to have the suggestions
from cl430, and the power consumption analysis in future.

So, I am abandoning the C++ features. I'll just make a Python script to
generate a header with namespaces and functions. I.e. it is just C, but
with namespaces.

	modified:   reg_funcs.hpp