Experimenting with namespaces of inlined functions
Instead of C++ features, just trying to structure a bunch of inline
functions in namespaces, according to the device structure.

It does not always inline even just functions or templated functions! It
looks like after a couple nested calls, cl430 gives up on inlining. But
it eventually worked for the bit field parameters, by passing the input
as the template parameter.

	modified:   README.md
	modified:   reg_funcs.hpp