An enum for the bitfield inputs
C style enum, to convert it to the int for the bitfield. Cpp does not
preserve type safety to some extend here: it does not compile if an
rvalue int is passed.

Ti compiler does capture a using definition from parent for come reason.
Although it does work in clang.

	modified:   bitlogic.hpp
	modified:   device_templates.hpp