Custom enum structs OptEnum for the bitfield values
The goal is to simplify how the user of BitField class specialises the
class and sets up the value options. It was expanding to too many lines:
a couple using statements (for cl430 compiler), then an enum with
options, and the definition of the options member function.

Now, the user just inherits a BitField, and an OPT type is provided with
it. The user only has to define constexpr static value options.

Unfortunately, it does not check the constexpr static options against
the width of the bitfield. It hits some problem and cl430 segfaults.
The same sort of assert does work in a test program with clang.

	modified:   bitlogic.hpp
	modified:   device_templates.hpp